app/requirements.txt
psycopg2-binary==2.8.4
environ-config==19.1.0